refactor: デ
